refactor: Update dependencies and components
Update npm dependencies to their latest stable versions:
- "@nextui-org/button": "^2.0.33"
- "@nextui-org/system": "^2.2.1"
- "@nextui-org/theme": "^2.2.5"
- "framer-motion": "^11.2.10"

In the AboutMe component, add useState and useEffect imports from React.

In the Navbar component, change the declaration of the 'links' variable from const to let.

In the tailwind.config.ts file, import '@nextui-org/theme' and add it as a plugin.

In the Project component, add an image prop and display the image using an img tag. Also, add a status prop and display it below the project description.

These changes improve the project's dependencies, component functionality, and styling.